linux/requirements.txt
requests == 2.31.0
google-cloud-monitoring == 2.16.0